Refactor executor
