+++ syncwriter.go
@@ -1,0 +1,17 @@
+package goyek
+
+import (
+	"io"
+	"sync"
+)
+
+type syncWriter struct {
+	io.Writer
+	mtx sync.Mutex
+}
+
+func (w *syncWriter) Write(p []byte) (int, error) {
+	defer func() { w.mtx.Unlock() }()
+	w.mtx.Lock()
+	return w.Writer.Write(p)
+}
